Redesigned how panels work
Made the panels full height of the screen, remove the sliding effect.
Took out that code. A lot of clean up in the calculations and the code
in general.

Remove a lot of the old styles, looks crappy now, but more functional.
Made some modifications to the html to reduce complexity.